Backend/SQL/requirements.txt
pika
mysql.connector